@@ -13,11 +13,10 @@ skip = ".git,.mypy_cache,.pytest_cache,.ruff_cache,.tox,.vscode,build,*/translat
 ignore-words-list = "hass"
 
 [tool.mypy]
-python_version = "3.12"
+python_version = "3.13"
 platform = "linux"
 show_error_codes = true
 follow_imports = "normal"
-enable_incomplete_feature = "NewGenericSyntax"
 local_partial_types = true
 strict_equality = true
 no_implicit_optional = true
@@ -75,7 +74,7 @@ exclude = [
 ]
 line-length = 88
 indent-width = 4
-target-version = "py311"
+target-version = "py313"
 
 [tool.ruff.lint]
 select = [
@@ -119,7 +118,6 @@ select = [
     "S317",  # suspicious-xml-sax-usage
     "S318",  # suspicious-xml-mini-dom-usage
     "S319",  # suspicious-xml-pull-dom-usage
-    "S320",  # suspicious-xmle-tree-usage
     "S601",  # paramiko-call
     "S602",  # subprocess-popen-with-shell-equals-true
     "S604",  # call-with-shell-equals-true
@@ -207,7 +205,7 @@ exclude_lines = [
 [tool.tox]
 legacy_tox_ini = """
     [tox]
-    envlist = py312, lint, type
+    envlist = py313, lint, type
     isolated_build = True
     skip_missing_interpreters = True
     ignore_basepython_conflict = True
@@ -215,7 +213,7 @@ legacy_tox_ini = """
 
     [gh-actions]
     python =
-        3.12: py312, lint, type
+        3.13: py313, lint, type
 
     [testenv]
     deps =
