@@ -2,85 +2,62 @@
 
 from __future__ import annotations
 
-from copy import copy
-from typing import Any, Final
+from dataclasses import fields, is_dataclass
+from typing import Any
 
+from homeassistant.components.diagnostics import async_redact_data
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.core import HomeAssistant
+from homeassistant.core import HomeAssistant, callback
 from pyplumio import __version__ as pyplumio_version
-from pyplumio.devices import Device
-from pyplumio.structures.product_info import ProductInfo
+from pyplumio.helpers.event_manager import EventManager
 
-from .const import ATTR_PASSWORD, ATTR_PRODUCT, CONF_HOST, CONF_UID
-
-REDACTED: Final = "**REDACTED**"
+from .connection import EcomaxConnection
+from .const import ATTR_PASSWORD, CONF_HOST, CONF_UID
 
 
-def _value_as_dict(value: Any) -> Any:
+@callback
+def _async_value_as_dict(value: Any) -> Any:
     """Return value as a dictionary."""
-    if isinstance(value, Device):
-        return dict(value.data)
+    if isinstance(value, EventManager):
+        return {**value.data}
 
-    if isinstance(value, ProductInfo):
-        return copy(value)
+    if is_dataclass(value):
+        return {field.name: getattr(value, field.name) for field in fields(value)}
 
     return value
 
 
-def _data_as_dict(data: dict[str, Any]) -> dict[str, Any]:
+@callback
+def _async_data_as_dict(data: dict[str, Any]) -> dict[str, Any]:
     """Return data as a dictionary."""
-    for key, value in data.items():
-        data[key] = (
-            _data_as_dict(dict(value))
+    items = {**data}
+
+    for key, value in items.items():
+        items[key] = (
+            _async_data_as_dict(value)
             if isinstance(value, dict)
-            else _value_as_dict(value)
+            else _async_value_as_dict(value)
         )
 
-    return data
-
-
-def _redact_device_data(data: dict[str, Any]) -> dict[str, Any]:
-    """Redact sensitive information from device data."""
-    sensitive: tuple[tuple[str, str | None], ...] = (
-        (ATTR_PRODUCT, "uid"),
-        (ATTR_PASSWORD, None),
-    )
-
-    for key, attribute in sensitive:
-        if key not in data:
-            continue
-
-        if attribute is None:
-            data[key] = REDACTED
-            continue
-
-        setattr(data[key], attribute, REDACTED)
-
-    return data
-
-
-def _redact_entry_data(data: dict[str, Any]) -> dict[str, Any]:
-    """Redact sensitive information from config entry data."""
-    for field in (CONF_UID, CONF_HOST):
-        if field in data:
-            data[field] = REDACTED
-
-    return data
+    return items
 
 
 async def async_get_config_entry_diagnostics(
     hass: HomeAssistant, entry: ConfigEntry
 ) -> dict[str, Any]:
     """Return diagnostics for a config entry."""
-    data = entry.runtime_data
-    device: Device = data.connection.device
+    connection: EcomaxConnection = entry.runtime_data.connection
     return {
         "entry": {
             "title": entry.title,
-            "data": _redact_entry_data(dict(entry.data)),
+            "data": async_redact_data(entry.data, to_redact={CONF_UID, CONF_HOST}),
+            "options": dict(entry.options),
         },
         "pyplumio": {
             "version": pyplumio_version,
         },
-        "data": _redact_device_data(_data_as_dict(dict(device.data))),
+        "data": async_redact_data(
+            _async_data_as_dict(connection.device.data),
+            to_redact={CONF_UID, ATTR_PASSWORD},
+        ),
     }
