@@ -2,28 +2,26 @@
 
 from __future__ import annotations
 
-from collections.abc import Generator, Iterable
 from dataclasses import dataclass
 import logging
-from typing import Any, Final
+from typing import Any, Final, cast
 
 from homeassistant.components.select import SelectEntity, SelectEntityDescription
 from homeassistant.const import STATE_OFF
-from homeassistant.core import HomeAssistant
+from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from pyplumio.const import ProductType
-from pyplumio.structures.modules import ConnectedModules
 
 from . import PlumEcomaxConfigEntry
 from .connection import EcomaxConnection
-from .const import ALL
 from .entity import (
-    DescriptorT,
     EcomaxEntity,
     EcomaxEntityDescription,
     MixerEntity,
-    SubDescriptorT,
     SubdeviceEntityDescription,
+    async_get_by_index,
+    async_get_by_modules,
+    async_get_by_product_type,
 )
 
 STATE_SUMMER: Final = "summer"
@@ -109,60 +107,29 @@ class MixerSelect(MixerEntity, EcomaxSelect):
         super().__init__(connection, description)
 
 
-def get_by_product_type(
-    product_type: ProductType,
-    descriptions: Iterable[DescriptorT],
-) -> Generator[DescriptorT, None, None]:
-    """Filter descriptions by the product type."""
-    for description in descriptions:
-        if (
-            description.product_types == ALL
-            or product_type in description.product_types
-        ):
-            yield description
-
-
-def get_by_modules(
-    connected_modules: ConnectedModules,
-    descriptions: Iterable[DescriptorT],
-) -> Generator[DescriptorT, None, None]:
-    """Filter descriptions by connected modules."""
-    for description in descriptions:
-        if getattr(connected_modules, description.module, None) is not None:
-            yield description
-
-
-def get_by_index(
-    index: int, descriptions: Iterable[SubDescriptorT]
-) -> Generator[SubDescriptorT, None, None]:
-    """Filter mixer/circuit descriptions by the index."""
-    index += 1
-    for description in descriptions:
-        if description.indexes == ALL or index in description.indexes:
-            yield description
-
-
+@callback
 def async_setup_ecomax_selects(connection: EcomaxConnection) -> list[EcomaxSelect]:
     """Set up the ecoMAX selects."""
     return [
         EcomaxSelect(connection, description)
-        for description in get_by_modules(
+        for description in async_get_by_modules(
             connection.device.modules,
-            get_by_product_type(connection.product_type, SELECT_TYPES),
+            async_get_by_product_type(connection.product_type, SELECT_TYPES),
         )
     ]
 
 
+@callback
 def async_setup_mixer_selects(connection: EcomaxConnection) -> list[MixerSelect]:
     """Set up the mixer selects."""
     return [
         MixerSelect(connection, description, index)
-        for index in connection.device.mixers
-        for description in get_by_index(
+        for index in cast(dict[int, Any], connection.device.mixers)
+        for description in async_get_by_index(
             index,
-            get_by_modules(
+            async_get_by_modules(
                 connection.device.modules,
-                get_by_product_type(connection.product_type, MIXER_SELECT_TYPES),
+                async_get_by_product_type(connection.product_type, MIXER_SELECT_TYPES),
             ),
         )
     ]
