@@ -2,9 +2,10 @@
 
 from __future__ import annotations
 
-import datetime as dt
+from dataclasses import dataclass
+import difflib
 import logging
-from typing import Any, Final
+from typing import Final, NotRequired, TypedDict, cast
 
 from homeassistant.const import ATTR_NAME, STATE_OFF, STATE_ON
 from homeassistant.core import (
@@ -22,44 +23,28 @@ from homeassistant.helpers.service import (
     SelectedEntities,
     async_extract_referenced_entity_ids,
 )
-from homeassistant.util.json import JsonObjectType
-from pyplumio.const import UnitOfMeasurement
-from pyplumio.devices import Device
-from pyplumio.helpers.parameter import Number, Parameter
-from pyplumio.helpers.schedule import (
-    STATE_DAY,
-    STATE_NIGHT,
-    TIME_FORMAT,
-    Schedule,
-    ScheduleDay,
-    start_of_day_dt,
-)
+from pyplumio.const import State, UnitOfMeasurement
+from pyplumio.devices import Device, VirtualDevice
+from pyplumio.parameters import Number, NumericType, Parameter
+from pyplumio.structures.product_info import ProductInfo
+from pyplumio.structures.schedules import Schedule, ScheduleDay
 import voluptuous as vol
 
 from .connection import DEFAULT_TIMEOUT, EcomaxConnection
-from .const import (
-    ATTR_END,
-    ATTR_PRESET,
-    ATTR_PRODUCT,
-    ATTR_SCHEDULES,
-    ATTR_START,
-    ATTR_TYPE,
-    ATTR_VALUE,
-    ATTR_WEEKDAYS,
-    DOMAIN,
-    WEEKDAYS,
-    DeviceType,
-)
+from .const import ATTR_PRODUCT, ATTR_VALUE, DOMAIN, VIRTUAL_DEVICES, WEEKDAYS
 
-SCHEDULES: Final = (
-    "heating",
-    "water_heater",
-)
+ATTR_START: Final = "start"
+ATTR_END: Final = "end"
+ATTR_PRESET: Final = "preset"
+ATTR_SCHEDULES: Final = "schedules"
+ATTR_TYPE: Final = "type"
+ATTR_WEEKDAYS: Final = "weekdays"
 
-PRESETS: Final = (
-    STATE_DAY,
-    STATE_NIGHT,
-)
+PRESET_DAY: Final = "day"
+PRESET_NIGHT: Final = "night"
+PRESETS = (PRESET_DAY, PRESET_NIGHT)
+
+SCHEDULES: Final = ("heating", "water_heater")
 
 SERVICE_GET_PARAMETER = "get_parameter"
 SERVICE_GET_PARAMETER_SCHEMA = make_entity_service_schema(
@@ -112,16 +97,17 @@ def async_extract_target_device(
             translation_placeholders={"device": device_id},
         )
 
+    ecomax = connection.device
     identifier = list(device.identifiers)[0][1]
-    for device_type in (DeviceType.MIXER, DeviceType.THERMOSTAT):
+    for device_type in VIRTUAL_DEVICES:
         if f"-{device_type}-" in identifier:
             index = int(identifier.split("-", 3).pop())
-            sub_devices: dict[int, Device] = connection.device.get_nowait(
-                f"{device_type}s", {}
+            devices = cast(
+                dict[int, VirtualDevice], ecomax.get_nowait(f"{device_type}s", {})
             )
-            return sub_devices.get(index, connection.device)
+            return devices.get(index, ecomax)
 
-    return connection.device
+    return ecomax
 
 
 @callback
@@ -142,48 +128,124 @@ def async_extract_referenced_devices(
     return devices
 
 
-async def async_get_device_parameter(
-    device: Device, name: str
-) -> dict[str, Any] | None:
-    """Get device parameter."""
-    try:
-        parameter = await device.get(name, timeout=DEFAULT_TIMEOUT)
-    except TimeoutError as e:
-        raise HomeAssistantError(
-            translation_domain=DOMAIN,
-            translation_key="get_parameter_timeout",
-            translation_placeholders={"parameter": name},
-        ) from e
+@dataclass
+class DeviceId:
+    """Represents a device id.
 
-    if not isinstance(parameter, Parameter):
-        raise ServiceValidationError(
-            translation_domain=DOMAIN,
-            translation_key="invalid_parameter",
-            translation_placeholders={"parameter": name},
-        )
+    This class contains information that can be used to identify
+    a specific device.
+    """
 
-    ecomax = device.parent if hasattr(device, "parent") else device
-    product = ecomax.get_nowait(ATTR_PRODUCT, default=None)
-    device_uid = product.uid if product is not None else "unknown"
+    type: str
+    index: int
+
+
+@dataclass
+class ProductId:
+    """Represents a product id.
+
+    This class contains information that can be used to identify
+    a product.
+    """
+
+    model: str
+    uid: str
+
+
+class ParameterResponse(TypedDict):
+    """Represents a response from get/set parameter services."""
+
+    name: str
+    value: NumericType | State | bool
+    min_value: NumericType | State | bool
+    max_value: NumericType | State | bool
+    step: NotRequired[float]
+    unit_of_measurement: NotRequired[str | None]
+    device: NotRequired[DeviceId]
+    product: NotRequired[ProductId]
+
+
+@callback
+def async_make_parameter_response(
+    device: Device, parameter: Parameter
+) -> ParameterResponse:
+    """Make a parameter response."""
+    response: ParameterResponse = {
+        "name": parameter.description.name,
+        "value": parameter.value,
+        "min_value": parameter.min_value,
+        "max_value": parameter.max_value,
+    }
+
     if isinstance(parameter, Number):
-        unit_of_measurement = (
+        response["step"] = parameter.description.step
+        response["unit_of_measurement"] = (
             parameter.unit_of_measurement.value
             if isinstance(parameter.unit_of_measurement, UnitOfMeasurement)
             else parameter.unit_of_measurement
         )
-    else:
-        unit_of_measurement = None
 
-    return {
-        "name": name,
-        "value": parameter.value,
-        "min_value": parameter.min_value,
-        "max_value": parameter.max_value,
-        "unit_of_measurement": unit_of_measurement,
-        "device_type": device.__class__.__name__.lower(),
-        "device_uid": device_uid,
-        "device_index": device.index + 1 if hasattr(device, "index") else 0,
-    }
+    if isinstance(device, VirtualDevice):
+        response["device"] = DeviceId(
+            type=device.__class__.__name__.lower(), index=device.index + 1
+        )
+        device = device.parent
+
+    if product_info := cast(ProductInfo | None, device.get_nowait(ATTR_PRODUCT)):
+        response["product"] = ProductId(model=product_info.model, uid=product_info.uid)
+
+    return response
+
+
+SUGGESTION_SCORE: Final = 0.6
+
+
+@callback
+def async_suggest_device_parameter_name(device: Device, name: str) -> str | None:
+    """Get the parameter name suggestion."""
+    parameter_names = [
+        name for name, value in device.data.items() if isinstance(value, Parameter)
+    ]
+    matches = difflib.get_close_matches(
+        name, parameter_names, n=1, cutoff=SUGGESTION_SCORE
+    )
+    return matches[0] if matches else None
+
+
+@callback
+def async_validate_device_parameter(device: Device, name: str) -> Parameter:
+    """Validate the device parameter."""
+    parameter = device.get_nowait(name, None)
+    if not parameter:
+        suggestion = async_suggest_device_parameter_name(device, name)
+        if suggestion:
+            raise HomeAssistantError(
+                translation_domain=DOMAIN,
+                translation_key="parameter_not_found_with_suggestion",
+                translation_placeholders={"parameter": name, "suggestion": suggestion},
+            )
+        else:
+            raise HomeAssistantError(
+                translation_domain=DOMAIN,
+                translation_key="parameter_not_found",
+                translation_placeholders={"parameter": name},
+            )
+
+    if not isinstance(parameter, Parameter):
+        raise ServiceValidationError(
+            translation_domain=DOMAIN,
+            translation_key="property_not_writable",
+            translation_placeholders={"property": name},
+        )
+
+    return parameter
+
+
+@callback
+def async_get_device_parameter(device: Device, name: str) -> ParameterResponse:
+    """Get device parameter."""
+    parameter = async_validate_device_parameter(device, name)
+    return async_make_parameter_response(device, parameter)
 
 
 @callback
@@ -199,16 +261,12 @@ def async_setup_get_parameter_service(
         name = service_call.data[ATTR_NAME]
         selected = async_extract_referenced_entity_ids(hass, service_call)
         devices = async_extract_referenced_devices(hass, connection, selected)
-
-        return {
+        response = {
             "parameters": [
-                parameter
-                for parameter in [
-                    await async_get_device_parameter(device, name) for device in devices
-                ]
-                if parameter is not None
+                async_get_device_parameter(device, name) for device in devices
             ]
         }
+        return cast(ServiceResponse, response)
 
     hass.services.async_register(
         DOMAIN,
@@ -219,17 +277,14 @@ def async_setup_get_parameter_service(
     )
 
 
-async def async_set_device_parameter(device: Device, name: str, value: float) -> bool:
+@callback
+def async_set_device_parameter(
+    device: Device, name: str, value: float
+) -> ParameterResponse:
     """Set device parameter."""
+    parameter = async_validate_device_parameter(device, name)
     try:
-        return await device.set(name, value, timeout=DEFAULT_TIMEOUT)
-    except TypeError as e:
-        raise ServiceValidationError(
-            str(e),
-            translation_domain=DOMAIN,
-            translation_key="invalid_parameter",
-            translation_placeholders={"parameter": name},
-        ) from e
+        parameter.set_nowait(value, timeout=DEFAULT_TIMEOUT)
     except ValueError as e:
         raise ServiceValidationError(
             str(e),
@@ -237,13 +292,8 @@ async def async_set_device_parameter(device: Device, name: str, value: float) ->
             translation_key="invalid_parameter_value",
             translation_placeholders={"parameter": name, "value": str(value)},
         ) from e
-    except TimeoutError as e:
-        raise HomeAssistantError(
-            str(e),
-            translation_domain=DOMAIN,
-            translation_key="set_parameter_timeout",
-            translation_placeholders={"parameter": name},
-        ) from e
+
+    return async_make_parameter_response(device, parameter)
 
 
 @callback
@@ -252,43 +302,28 @@ def async_setup_set_parameter_service(
 ) -> None:
     """Set up the service to set a device parameter."""
 
-    async def _async_set_parameter_service(service_call: ServiceCall) -> None:
+    async def _async_set_parameter_service(
+        service_call: ServiceCall,
+    ) -> ServiceResponse | None:
         """Service to set a device parameter."""
         name = service_call.data[ATTR_NAME]
         value = service_call.data[ATTR_VALUE]
         selected = async_extract_referenced_entity_ids(hass, service_call)
         devices = async_extract_referenced_devices(hass, connection, selected)
-
-        if not any(
-            {
-                await async_set_device_parameter(device, name, value)
-                for device in devices
-            }
-        ):
-            raise HomeAssistantError(
-                translation_domain=DOMAIN,
-                translation_key="set_parameter_failed",
-                translation_placeholders={"parameter": name},
-            )
+        response = {
+            "parameters": [
+                async_set_device_parameter(device, name, value) for device in devices
+            ]
+        }
+        return cast(ServiceResponse, response) if service_call.return_response else None
 
     hass.services.async_register(
         DOMAIN,
         SERVICE_SET_PARAMETER,
         _async_set_parameter_service,
         schema=SERVICE_SET_PARAMETER_SCHEMA,
+        supports_response=SupportsResponse.OPTIONAL,
     )
-
-
-def async_get_schedule_day_data(
-    schedule_day: ScheduleDay,
-) -> JsonObjectType:
-    """Format the schedule day as a dictionary."""
-    return {
-        (start_of_day_dt + dt.timedelta(minutes=30 * index)).strftime(TIME_FORMAT): (
-            STATE_DAY if value else STATE_NIGHT
-        )
-        for index, value in enumerate(schedule_day.intervals)
-    }
 
 
 @callback
@@ -313,7 +348,10 @@ def async_setup_get_schedule_service(
         schedule = schedules[schedule_type]
         return {
             "schedule": {
-                weekday: async_get_schedule_day_data(getattr(schedule, weekday))
+                weekday: {
+                    interval: PRESET_DAY if state == STATE_ON else PRESET_NIGHT
+                    for interval, state in getattr(schedule, weekday).items()
+                }
                 for weekday in weekdays
             },
         }
@@ -340,8 +378,9 @@ def async_setup_set_schedule_service(
         preset = service_call.data[ATTR_PRESET]
         start_time = service_call.data[ATTR_START]
         end_time = service_call.data[ATTR_END]
-
-        schedules = connection.device.get_nowait(ATTR_SCHEDULES, {})
+        schedules = cast(
+            dict[str, Schedule], connection.device.get_nowait(ATTR_SCHEDULES, {})
+        )
         if schedule_type not in schedules:
             raise ServiceValidationError(
                 translation_domain=DOMAIN,
@@ -349,11 +388,12 @@ def async_setup_set_schedule_service(
                 translation_placeholders={"schedule": schedule_type},
             )
 
-        schedule: Schedule = schedules[schedule_type]
+        schedule = schedules[schedule_type]
+        state = bool(preset == "day")
         for weekday in weekdays:
             schedule_day: ScheduleDay = getattr(schedule, weekday)
             try:
-                schedule_day.set_state(preset, start_time[:-3], end_time[:-3])
+                schedule_day.set_state(state, start_time[:-3], end_time[:-3])
             except ValueError as e:
                 raise ServiceValidationError(
                     str(e),
