@@ -7,27 +7,35 @@ from homeassistant.components.switch import (
     SERVICE_TURN_OFF,
     SERVICE_TURN_ON,
 )
-from homeassistant.const import ATTR_ENTITY_ID, ATTR_FRIENDLY_NAME, STATE_OFF, STATE_ON
+from homeassistant.const import (
+    ATTR_ENTITY_ID,
+    ATTR_FRIENDLY_NAME,
+    STATE_OFF,
+    STATE_ON,
+    Platform,
+)
 from homeassistant.core import HomeAssistant, State
 from homeassistant.helpers import entity_registry as er
-from pyplumio.helpers.parameter import ParameterValues
-from pyplumio.structures.ecomax_parameters import (
-    ATTR_ECOMAX_CONTROL,
+from pyplumio.parameters import ParameterValues
+from pyplumio.parameters.ecomax import (
     EcomaxNumber,
     EcomaxNumberDescription,
     EcomaxSwitch,
     EcomaxSwitchDescription,
 )
-from pyplumio.structures.mixer_parameters import (
+from pyplumio.parameters.mixer import (
     MixerNumber,
     MixerNumberDescription,
     MixerSwitch,
     MixerSwitchDescription,
 )
+from pyplumio.structures.ecomax_parameters import ATTR_ECOMAX_CONTROL
 import pytest
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
 from custom_components.plum_ecomax.connection import EcomaxConnection
+from custom_components.plum_ecomax.const import ATTR_ENTITIES
+from tests.conftest import dispatch_value
 
 
 @pytest.fixture(autouse=True)
@@ -107,7 +115,8 @@ async def test_ecomax_control_switch(
     assert state.attributes[ATTR_FRIENDLY_NAME] == "ecoMAX Controller switch"
 
     # Dispatch new value.
-    await connection.device.dispatch(
+    await dispatch_value(
+        connection.device,
         ATTR_ECOMAX_CONTROL,
         EcomaxSwitch(
             device=connection.device,
@@ -168,7 +177,8 @@ async def test_water_heater_disinfection_switch(
     )
 
     # Dispatch new value.
-    await connection.device.dispatch(
+    await dispatch_value(
+        connection.device,
         water_heater_disinfection_switch_key,
         EcomaxSwitch(
             device=connection.device,
@@ -228,7 +238,8 @@ async def test_water_heater_pump_switch(
     assert state.attributes[ATTR_FRIENDLY_NAME] == "ecoMAX Water heater pump switch"
 
     # Dispatch new value.
-    await connection.device.dispatch(
+    await dispatch_value(
+        connection.device,
         water_heater_pump_switch_key,
         EcomaxNumber(
             device=connection.device,
@@ -284,7 +295,8 @@ async def test_weather_control_switch(
     assert state.attributes[ATTR_FRIENDLY_NAME] == "ecoMAX Weather control switch"
 
     # Dispatch new value.
-    await connection.device.dispatch(
+    await dispatch_value(
+        connection.device,
         weather_control_switch_key,
         EcomaxSwitch(
             device=connection.device,
@@ -340,7 +352,8 @@ async def test_fuzzy_logic_switch(
     assert state.attributes[ATTR_FRIENDLY_NAME] == "ecoMAX Fuzzy logic switch"
 
     # Dispatch new value.
-    await connection.device.dispatch(
+    await dispatch_value(
+        connection.device,
         fuzzy_logic_switch_key,
         EcomaxSwitch(
             device=connection.device,
@@ -396,7 +409,8 @@ async def test_heating_schedule_switch(
     assert state.attributes[ATTR_FRIENDLY_NAME] == "ecoMAX Heating schedule switch"
 
     # Dispatch new value.
-    await connection.device.dispatch(
+    await dispatch_value(
+        connection.device,
         heating_schedule_switch_key,
         EcomaxSwitch(
             device=connection.device,
@@ -454,7 +468,8 @@ async def test_water_heater_schedule_switch(
     assert state.attributes[ATTR_FRIENDLY_NAME] == "ecoMAX Water heater schedule switch"
 
     # Dispatch new value.
-    await connection.device.dispatch(
+    await dispatch_value(
+        connection.device,
         water_heater_schedule_switch_key,
         EcomaxSwitch(
             device=connection.device,
@@ -794,3 +809,99 @@ async def test_circuit_enable_circuit_switch_is_unavailable_for_second_circuit(
     entity_registry = er.async_get(hass)
     entry = entity_registry.async_get(enable_circuit_entity_id)
     assert not entry
+
+
+@pytest.mark.parametrize(
+    ("source_device", "entity_id", "friendly_name"),
+    (
+        (
+            "ecomax",
+            "switch.ecomax_test_custom_switch",
+            "ecoMAX Test custom switch",
+        ),
+        (
+            "mixer_0",
+            "switch.ecomax_mixer_1_test_custom_switch",
+            "ecoMAX Mixer 1 Test custom switch",
+        ),
+        (
+            "mixer_1",
+            "switch.ecomax_mixer_2_test_custom_switch",
+            "ecoMAX Mixer 2 Test custom switch",
+        ),
+        (
+            "thermostat_0",
+            "switch.ecomax_thermostat_1_test_custom_switch",
+            "ecoMAX Thermostat 1 Test custom switch",
+        ),
+    ),
+)
+@pytest.mark.usefixtures("ecomax_p", "mixers", "thermostats", "custom_fields")
+async def test_custom_switches(
+    source_device: str,
+    entity_id: str,
+    friendly_name: str,
+    hass: HomeAssistant,
+    connection: EcomaxConnection,
+    config_entry: MockConfigEntry,
+    setup_integration,
+    async_turn_off,
+    async_turn_on,
+) -> None:
+    """Test custom switches."""
+    custom_switch_key = "custom_switch"
+    await setup_integration(
+        hass,
+        config_entry,
+        options={
+            ATTR_ENTITIES: {
+                Platform.SWITCH: {
+                    custom_switch_key: {
+                        "name": "Test custom switch",
+                        "key": custom_switch_key,
+                        "source_device": source_device,
+                    }
+                }
+            }
+        },
+    )
+
+    # Test entry.
+    entity_registry = er.async_get(hass)
+    entry = entity_registry.async_get(entity_id)
+    assert entry
+
+    # Get initial state.
+    state = hass.states.get(entity_id)
+    assert isinstance(state, State)
+    assert state.state == STATE_OFF
+    assert state.attributes[ATTR_FRIENDLY_NAME] == friendly_name
+
+    # Dispatch new state.
+    new_state = EcomaxSwitch(
+        device=connection.device,
+        values=ParameterValues(value=1, min_value=0, max_value=1),
+        description=EcomaxSwitchDescription(custom_switch_key),
+    )
+    await dispatch_value(
+        connection.device, custom_switch_key, new_state, source_device=source_device
+    )
+    state = hass.states.get(entity_id)
+    assert isinstance(state, State)
+    assert state.state == STATE_ON
+
+    # Turn off.
+    with patch("pyplumio.devices.Device.set_nowait") as mock_set_nowait:
+        state = await async_turn_off(hass, entity_id)
+
+    mock_set_nowait.assert_called_once_with(custom_switch_key, STATE_OFF)
+    assert isinstance(state, State)
+    assert state.state == STATE_OFF
+
+    # Turn on.
+    with patch("pyplumio.devices.Device.set_nowait") as mock_set_nowait:
+        state = await async_turn_on(hass, entity_id)
+
+    mock_set_nowait.assert_called_once_with(custom_switch_key, STATE_ON)
+    assert isinstance(state, State)
+    assert state.state == STATE_ON
