@@ -24,8 +24,8 @@ from homeassistant.const import ATTR_ENTITY_ID, ATTR_FRIENDLY_NAME, STATE_OFF
 from homeassistant.core import HomeAssistant, State
 from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.entity_registry import RegistryEntry
-from pyplumio.helpers.parameter import ParameterValues
-from pyplumio.structures.ecomax_parameters import EcomaxNumber, EcomaxNumberDescription
+from pyplumio.parameters import ParameterValues
+from pyplumio.parameters.ecomax import EcomaxNumber, EcomaxNumberDescription
 import pytest
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
